+++ loadtest/k6/scen5.js
@@ -5,54 +5,52 @@ import { sleep } from 'k6';
 const host='kong-proxy.kong:8443';
 //const host='35.241.175.116';
 
-
 export const options = {
   insecureSkipTLSVerify: true,
-  scenarios: {    
+  scenarios: {
     scenOk: {
-      exec: 'scen2endurance',
-      
+      exec: 'scen5',
       /*executor: 'per-vu-iterations',
       vus: 1,
       iterations: 1,
-      maxDuration: '1s'*/
+      maxDuration: '1s',*/
       
       executor: 'ramping-vus',
-      startvus: 0,
+      startvus: 2,
       stages: [
-        { duration: '30s', target: 6 },
-        { duration: '30s', target: 12 },
-        { duration: '30s', target: 20 },
-        { duration: '28800s', target: 20 },
+        { duration: '300s', target: 8 },
+        { duration: '300s', target: 12 },
+        { duration: '900s', target: 16 },
       ],
       gracefulRampDown: '5s',
     },
   },
 };
 
-export function scen2endurance () {
+export function scen5 () {
   const calcReq = 
   `<?xml version=\"1.0\" encoding=\"utf-8\"?>
   <soap:Envelope xmlns:xsi=\"http://www.w3.org/2001/XMLSchema-instance\" xmlns:xsd=\"http://www.w3.org/2001/XMLSchema\" xmlns:soap=\"http://schemas.xmlsoap.org/soap/envelope/\">
     <soap:Body>
-      <Add xmlns=\"http://tempuri.org/\">
+      <Subtract xmlns=\"http://tempuri.org/\">
         <intA>5</intA>
-        <intB>7</intB>
-      </Add>
+        <intB>0</intB>
+      </Subtract>
     </soap:Body>
   </soap:Envelope>`;
   
-  const result = http.post('https://'+host+'/scen2endurance', calcReq, {
+  const result = http.post('https://'+host+'/scen5', calcReq, {
     headers: { 
         'Content-Type': 'text/xml; charset=utf-8',
+        'SOAPAction': 'http://tempuri.org/Add'
       },
   });
 
   check(result, {
     'http response status code is 200': result.status === 200,
     'Content-Type': result.headers['Content-Type'] === 'text/xml;charset=utf-8',
+    'X-Soap-Region': result.headers['X-Soap-Region'] === 'soap2',
     "calculator Result": result =>
-      result.body.includes("<AddResult>12</AddResult>"),
+      result.body.includes("<KongResult>15</KongResult>"),
   });
-  sleep(0.01);
 }
